+++ src/ZEO/asyncio/tests.py
@@ -7,17 +7,62 @@ import asyncio
 import collections
 import logging
 import pdb
-import pickle
 import struct
 import unittest
-import ZEO.Exceptions
+
+from ..Exceptions import ClientDisconnected, ProtocolError
+from ..ClientStorage import m64
 
 from .testing import Loop
 from .client import ClientRunner, Fallback
-from ..Exceptions import ClientDisconnected
-from ..ClientStorage import m64
+from .server import new_connection, best_protocol_version
+from .marshal import encoder, decode
 
-class AsyncTests(setupstack.TestCase, ClientRunner):
+class Base(object):
+
+    def setUp(self):
+        super(Base, self).setUp()
+        self.encode = encoder()
+
+    def unsized(self, data, unpickle=False):
+        result = []
+        while data:
+            size, message, *data = data
+            self.assertEqual(struct.unpack(">I", size)[0], len(message))
+            if unpickle:
+                message = decode(message)
+            result.append(message)
+
+        if len(result) == 1:
+            result = result[0]
+        return result
+
+    def parse(self, data):
+        return self.unsized(data, True)
+
+    target = None
+    def send(self, method, *args, **kw):
+        target = kw.pop('target', self.target)
+        called = kw.pop('called', True)
+        no_output = kw.pop('no_output', True)
+        self.assertFalse(kw)
+
+        self.loop.protocol.data_received(
+            sized(self.encode(0, True, method, args)))
+        if target is not None:
+            target = getattr(target, method)
+            if called:
+                target.assert_called_with(*args)
+                target.reset_mock()
+            else:
+                self.assertFalse(target.called)
+        if no_output:
+            self.assertFalse(self.loop.transport.pop())
+
+    def pop(self, count=None, parse=True):
+        return self.unsized(self.loop.transport.pop(count), parse)
+
+class ClientTests(Base, setupstack.TestCase, ClientRunner):
 
     def start(self,
               addrs=(('127.0.0.1', 8200), ), loop_addrs=None,
@@ -28,6 +73,7 @@ class AsyncTests(setupstack.TestCase, ClientRunner):
         # object and a cache.
 
         wrapper = mock.Mock()
+        self.target = wrapper
         cache = MemoryCache()
         self.set_options(addrs, wrapper, cache, 'TEST', read_only, timeout=1)
 
@@ -39,42 +85,35 @@ class AsyncTests(setupstack.TestCase, ClientRunner):
         protocol = loop.protocol
         transport = loop.transport
 
-        def send(meth, *args):
-            loop.protocol.data_received(
-                sized(pickle.dumps((0, True, meth, args), 3)))
-
-        def respond(message_id, result):
-            loop.protocol.data_received(
-                sized(pickle.dumps((message_id, False, '.reply', result), 3)))
-
         if finish_start:
             protocol.data_received(sized(b'Z3101'))
-            self.assertEqual(self.unsized(transport.pop(2)), b'Z3101')
-            parse = self.parse
-            self.assertEqual(parse(transport.pop()),
+            self.assertEqual(self.pop(2, False), b'Z3101')
+            self.assertEqual(self.pop(),
                              [(1, False, 'register', ('TEST', False)),
                               (2, False, 'lastTransaction', ()),
                               ])
-            respond(1, None)
-            respond(2, 'a'*8)
-            self.assertEqual(parse(transport.pop()), (3, False, 'get_info', ()))
-            respond(3, dict(length=42))
+            self.respond(1, None)
+            self.respond(2, 'a'*8)
+            self.assertEqual(self.pop(), (3, False, 'get_info', ()))
+            self.respond(3, dict(length=42))
 
-        return (wrapper, cache, self.loop, self.client, protocol, transport,
-                send, respond)
+        return (wrapper, cache, self.loop, self.client, protocol, transport)
+
+    def respond(self, message_id, result):
+        self.loop.protocol.data_received(
+            sized(self.encode(message_id, False, '.reply', result)))
 
     def wait_for_result(self, future, timeout):
         return future
 
-    def testBasics(self):
+    def testClientBasics(self):
 
         # Here, we'll go through the basic usage of the asyncio ZEO
         # network client.  The client is responsible for the core
         # functionality of a ZEO client storage.  The client storage
         # is largely just a wrapper around the asyncio client.
 
-        wrapper, cache, loop, client, protocol, transport, send, respond = (
-            self.start())
+        wrapper, cache, loop, client, protocol, transport = self.start()
         self.assertFalse(wrapper.notify_disconnected.called)
 
         # The client isn't connected until the server sends it some data.
@@ -87,9 +126,8 @@ class AsyncTests(setupstack.TestCase, ClientRunner):
 
         # The client sends back a handshake, and registers the
         # storage, and requests the last transaction.
-        self.assertEqual(self.unsized(transport.pop(2)), b'Z3101')
-        parse = self.parse
-        self.assertEqual(parse(transport.pop()),
+        self.assertEqual(self.pop(2, False), b'Z5')
+        self.assertEqual(self.pop(),
                          [(1, False, 'register', ('TEST', False)),
                           (2, False, 'lastTransaction', ()),
                           ])
@@ -119,37 +157,36 @@ class AsyncTests(setupstack.TestCase, ClientRunner):
 
         # Let's respond to those first 2 calls:
 
-        respond(1, None)
-        respond(2, 'a'*8)
+        self.respond(1, None)
+        self.respond(2, 'a'*8)
 
         # After verification, the client requests info:
-        self.assertEqual(parse(transport.pop()), (3, False, 'get_info', ()))
-        respond(3, dict(length=42))
+        self.assertEqual(self.pop(), (3, False, 'get_info', ()))
+        self.respond(3, dict(length=42))
 
         # Now we're connected, the cache was initialized, and the
         # queued message has been sent:
         self.assert_(client.connected.done())
         self.assertEqual(cache.getLastTid(), 'a'*8)
-        self.assertEqual(parse(transport.pop()), (4, False, 'foo', (1, 2)))
+        self.assertEqual(self.pop(), (4, False, 'foo', (1, 2)))
 
         # The wrapper object (ClientStorage) has been notified:
         wrapper.notify_connected.assert_called_with(client, {'length': 42})
 
-        respond(4, 42)
+        self.respond(4, 42)
         self.assertEqual(f1.result(), 42)
 
         # Now we can make async calls:
         f2 = self.async('bar', 3, 4)
         self.assert_(f2.done() and f2.exception() is None)
-        self.assertEqual(parse(transport.pop()), (0, True, 'bar', (3, 4)))
+        self.assertEqual(self.pop(), (0, True, 'bar', (3, 4)))
 
         # Loading objects gets special handling to leverage the cache.
         loaded = self.load_before(b'1'*8, m64)
 
         # The data wasn't in the cache, so we make a server call:
-        self.assertEqual(parse(transport.pop()),
-                         (5, False, 'loadBefore', (b'1'*8, m64)))
-        respond(5, (b'data', b'a'*8, None))
+        self.assertEqual(self.pop(), (5, False, 'loadBefore', (b'1'*8, m64)))
+        self.respond(5, (b'data', b'a'*8, None))
         self.assertEqual(loaded.result(), (b'data', b'a'*8, None))
 
         # If we make another request, it will be satisfied from the cache:
@@ -158,15 +195,12 @@ class AsyncTests(setupstack.TestCase, ClientRunner):
         self.assertFalse(transport.data)
 
         # Let's send an invalidation:
-        send('invalidateTransaction', b'b'*8, [b'1'*8])
-        wrapper.invalidateTransaction.assert_called_with(b'b'*8, [b'1'*8])
-        wrapper.invalidateTransaction.reset_mock()
+        self.send('invalidateTransaction', b'b'*8, [b'1'*8])
 
         # Now, if we try to load current again, we'll make a server request.
         loaded = self.load_before(b'1'*8, m64)
-        self.assertEqual(parse(transport.pop()),
-                         (6, False, 'loadBefore', (b'1'*8, m64)))
-        respond(6, (b'data2', b'b'*8, None))
+        self.assertEqual(self.pop(), (6, False, 'loadBefore', (b'1'*8, m64)))
+        self.respond(6, (b'data2', b'b'*8, None))
         self.assertEqual(loaded.result(), (b'data2', b'b'*8, None))
 
         # Loading non-current data may also be satisfied from cache
@@ -178,9 +212,9 @@ class AsyncTests(setupstack.TestCase, ClientRunner):
         self.assertFalse(transport.data)
         loaded = self.load_before(b'1'*8, b'_'*8)
 
-        self.assertEqual(parse(transport.pop()),
+        self.assertEqual(self.pop(),
                          (7, False, 'loadBefore', (b'1'*8, b'_'*8)))
-        respond(7, (b'data0', b'^'*8, b'_'*8))
+        self.respond(7, (b'data0', b'^'*8, b'_'*8))
         self.assertEqual(loaded.result(), (b'data0', b'^'*8, b'_'*8))
 
         # When committing transactions, we need to update the cache
@@ -202,9 +236,9 @@ class AsyncTests(setupstack.TestCase, ClientRunner):
                          cache.load(b'2'*8) or
                          cache.load(b'4'*8))
         self.assertEqual(cache.load(b'1'*8), (b'data2', b'b'*8))
-        self.assertEqual(parse(transport.pop()),
+        self.assertEqual(self.pop(),
                          (8, False, 'tpc_finish', (b'd'*8,)))
-        respond(8, b'e'*8)
+        self.respond(8, b'e'*8)
         self.assertEqual(committed.result(), b'e'*8)
         self.assertEqual(cache.load(b'1'*8), None)
         self.assertEqual(cache.load(b'2'*8), ('committed 2', b'e'*8))
@@ -216,9 +250,8 @@ class AsyncTests(setupstack.TestCase, ClientRunner):
         loaded = self.load_before(b'1'*8, m64)
         f1 = self.call('foo', 1, 2)
         self.assertFalse(loaded.done() or f1.done())
-        self.assertEqual(parse(transport.pop()),
-                         [(9, False, 'loadBefore', (b'1'*8, m64)),
-                          (10, False, 'foo', (1, 2))],
+        self.assertEqual(self.pop(), [(9, False, 'loadBefore', (b'1'*8, m64)),
+                                      (10, False, 'foo', (1, 2))],
                          )
         exc = TypeError(43)
 
@@ -246,15 +279,15 @@ class AsyncTests(setupstack.TestCase, ClientRunner):
         # protocol:
         protocol.data_received(sized(b'Z310'))
         self.assertEqual(self.unsized(transport.pop(2)), b'Z310')
-        self.assertEqual(parse(transport.pop()),
+        self.assertEqual(self.pop(),
                          [(1, False, 'register', ('TEST', False)),
                           (2, False, 'lastTransaction', ()),
                           ])
         self.assertFalse(wrapper.notify_connected.called)
-        respond(1, None)
-        respond(2, b'e'*8)
-        self.assertEqual(parse(transport.pop()), (3, False, 'get_info', ()))
-        respond(3, dict(length=42))
+        self.respond(1, None)
+        self.respond(2, b'e'*8)
+        self.assertEqual(self.pop(), (3, False, 'get_info', ()))
+        self.respond(3, dict(length=42))
 
         # Because the server tid matches the cache tid, we're done connecting
         wrapper.notify_connected.assert_called_with(client, {'length': 42})
@@ -274,8 +307,7 @@ class AsyncTests(setupstack.TestCase, ClientRunner):
         self.assertEqual(loop.transport, transport)
 
     def test_cache_behind(self):
-        wrapper, cache, loop, client, protocol, transport, send, respond = (
-            self.start())
+        wrapper, cache, loop, client, protocol, transport = self.start()
 
         cache.setLastTid(b'a'*8)
         cache.store(b'4'*8, b'a'*8, None, '4 data')
@@ -284,22 +316,20 @@ class AsyncTests(setupstack.TestCase, ClientRunner):
         self.assertFalse(client.connected.done() or transport.data)
         protocol.data_received(sized(b'Z3101'))
         self.assertEqual(self.unsized(transport.pop(2)), b'Z3101')
-        self.assertEqual(self.parse(transport.pop()),
+        self.assertEqual(self.pop(),
                          [(1, False, 'register', ('TEST', False)),
                           (2, False, 'lastTransaction', ()),
                           ])
-        respond(1, None)
-        respond(2, b'e'*8)
+        self.respond(1, None)
+        self.respond(2, b'e'*8)
 
         # We have to verify the cache, so we're not done connecting:
         self.assertFalse(client.connected.done())
-        self.assertEqual(self.parse(transport.pop()),
-                         (3, False, 'getInvalidations', (b'a'*8, )))
-        respond(3, (b'e'*8, [b'4'*8]))
+        self.assertEqual(self.pop(), (3, False, 'getInvalidations', (b'a'*8, )))
+        self.respond(3, (b'e'*8, [b'4'*8]))
 
-        self.assertEqual(self.parse(transport.pop()),
-                         (4, False, 'get_info', ()))
-        respond(4, dict(length=42))
+        self.assertEqual(self.pop(), (4, False, 'get_info', ()))
+        self.respond(4, dict(length=42))
 
         # Now that verification is done, we're done connecting
         self.assert_(client.connected.done() and not transport.data)
@@ -315,8 +345,7 @@ class AsyncTests(setupstack.TestCase, ClientRunner):
         self.assertFalse(wrapper.invalidateCache.called)
 
     def test_cache_way_behind(self):
-        wrapper, cache, loop, client, protocol, transport, send, respond = (
-            self.start())
+        wrapper, cache, loop, client, protocol, transport = self.start()
 
         cache.setLastTid(b'a'*8)
         cache.store(b'4'*8, b'a'*8, None, '4 data')
@@ -325,24 +354,22 @@ class AsyncTests(setupstack.TestCase, ClientRunner):
         self.assertFalse(client.connected.done() or transport.data)
         protocol.data_received(sized(b'Z3101'))
         self.assertEqual(self.unsized(transport.pop(2)), b'Z3101')
-        self.assertEqual(self.parse(transport.pop()),
+        self.assertEqual(self.pop(),
                          [(1, False, 'register', ('TEST', False)),
                           (2, False, 'lastTransaction', ()),
                           ])
-        respond(1, None)
-        respond(2, b'e'*8)
+        self.respond(1, None)
+        self.respond(2, b'e'*8)
 
         # We have to verify the cache, so we're not done connecting:
         self.assertFalse(client.connected.done())
-        self.assertEqual(self.parse(transport.pop()),
-                         (3, False, 'getInvalidations', (b'a'*8, )))
+        self.assertEqual(self.pop(), (3, False, 'getInvalidations', (b'a'*8, )))
 
         # We respond None, indicating that we're too far out of date:
-        respond(3, None)
+        self.respond(3, None)
 
-        self.assertEqual(self.parse(transport.pop()),
-                         (4, False, 'get_info', ()))
-        respond(4, dict(length=42))
+        self.assertEqual(self.pop(), (4, False, 'get_info', ()))
+        self.respond(4, dict(length=42))
 
         # Now that verification is done, we're done connecting
         self.assert_(client.connected.done() and not transport.data)
@@ -355,8 +382,8 @@ class AsyncTests(setupstack.TestCase, ClientRunner):
     def test_multiple_addresses(self):
         # We can pass multiple addresses to client constructor
         addrs = [('1.2.3.4', 8200), ('2.2.3.4', 8200)]
-        wrapper, cache, loop, client, protocol, transport, send, respond = (
-            self.start(addrs, ()))
+        wrapper, cache, loop, client, protocol, transport = self.start(
+            addrs, ())
 
         # We haven't connected yet
         self.assert_(protocol is None and transport is None)
@@ -381,7 +408,7 @@ class AsyncTests(setupstack.TestCase, ClientRunner):
         transport = loop.transport
         protocol.data_received(sized(b'Z3101'))
         self.assertEqual(self.unsized(transport.pop(2)), b'Z3101')
-        respond(1, None)
+        self.respond(1, None)
 
         # Now, when the first connection fails, it won't be retried,
         # because we're already connected.
@@ -394,19 +421,17 @@ class AsyncTests(setupstack.TestCase, ClientRunner):
     def test_bad_server_tid(self):
         # If in verification we get a server_tid behing the cache's, make sure
         # we retry the connection later.
-        wrapper, cache, loop, client, protocol, transport, send, respond = (
-            self.start())
+        wrapper, cache, loop, client, protocol, transport = self.start()
         cache.store(b'4'*8, b'a'*8, None, '4 data')
         cache.setLastTid('b'*8)
         protocol.data_received(sized(b'Z3101'))
         self.assertEqual(self.unsized(transport.pop(2)), b'Z3101')
-        parse = self.parse
-        self.assertEqual(parse(transport.pop()),
+        self.assertEqual(self.pop(),
                          [(1, False, 'register', ('TEST', False)),
                           (2, False, 'lastTransaction', ()),
                           ])
-        respond(1, None)
-        respond(2, 'a'*8)
+        self.respond(1, None)
+        self.respond(2, 'a'*8)
         self.assertFalse(client.connected.done() or transport.data)
         delay, func, args, _ = loop.later.pop(1) # first in later is heartbeat
         self.assert_(8 < delay < 10)
@@ -418,21 +443,21 @@ class AsyncTests(setupstack.TestCase, ClientRunner):
         transport = loop.transport
         protocol.data_received(sized(b'Z3101'))
         self.assertEqual(self.unsized(transport.pop(2)), b'Z3101')
-        self.assertEqual(parse(transport.pop()),
+        self.assertEqual(self.pop(),
                          [(1, False, 'register', ('TEST', False)),
                           (2, False, 'lastTransaction', ()),
                           ])
-        respond(1, None)
-        respond(2, 'b'*8)
-        self.assertEqual(parse(transport.pop()), (3, False, 'get_info', ()))
-        respond(3, dict(length=42))
+        self.respond(1, None)
+        self.respond(2, 'b'*8)
+        self.assertEqual(self.pop(), (3, False, 'get_info', ()))
+        self.respond(3, dict(length=42))
         self.assert_(client.connected.done() and not transport.data)
         self.assert_(client.ready)
 
     def test_readonly_fallback(self):
         addrs = [('1.2.3.4', 8200), ('2.2.3.4', 8200)]
-        wrapper, cache, loop, client, protocol, transport, send, respond = (
-            self.start(addrs, (), read_only=Fallback))
+        wrapper, cache, loop, client, protocol, transport = self.start(
+            addrs, (), read_only=Fallback)
 
         self.assertTrue(self.is_read_only())
 
@@ -442,20 +467,20 @@ class AsyncTests(setupstack.TestCase, ClientRunner):
         protocol.data_received(sized(b'Z3101'))
         self.assertEqual(self.unsized(transport.pop(2)), b'Z3101')
         # We see that the client tried a writable connection:
-        self.assertEqual(self.parse(transport.pop()),
+        self.assertEqual(self.pop(),
                          (1, False, 'register', ('TEST', False)))
         # We respond with a read-only exception:
-        respond(1, (ReadOnlyError, ReadOnlyError()))
+        self.respond(1, (ReadOnlyError, ReadOnlyError()))
         self.assertTrue(self.is_read_only())
 
         # The client tries for a read-only connection:
-        self.assertEqual(self.parse(transport.pop()),
+        self.assertEqual(self.pop(),
                          [(2, False, 'register', ('TEST', True)),
                           (3, False, 'lastTransaction', ()),
                           ])
         # We respond with successfully:
-        respond(2, None)
-        respond(3, 'b'*8)
+        self.respond(2, None)
+        self.respond(3, 'b'*8)
         self.assertTrue(self.is_read_only())
 
         # At this point, the client is ready and using the protocol,
@@ -466,9 +491,8 @@ class AsyncTests(setupstack.TestCase, ClientRunner):
         connected = client.connected
 
         # The client asks for info, and we respond:
-        self.assertEqual(self.parse(transport.pop()),
-                         (4, False, 'get_info', ()))
-        respond(4, dict(length=42))
+        self.assertEqual(self.pop(), (4, False, 'get_info', ()))
+        self.respond(4, dict(length=42))
 
         self.assert_(connected.done())
 
@@ -481,7 +505,7 @@ class AsyncTests(setupstack.TestCase, ClientRunner):
         self.assertTrue(self.is_read_only())
 
         # We respond and the writable connection succeeds:
-        respond(1, None)
+        self.respond(1, None)
         self.assertFalse(self.is_read_only())
 
         # at this point, a lastTransaction request is emitted:
@@ -501,28 +525,25 @@ class AsyncTests(setupstack.TestCase, ClientRunner):
         self.assertEqual(protocol.read_only, False)
 
         # Now, we finish verification
-        respond(2, 'b'*8)
-        respond(3, dict(length=42))
+        self.respond(2, 'b'*8)
+        self.respond(3, dict(length=42))
         self.assert_(client.ready)
         self.assert_(client.connected.done())
 
     def test_invalidations_while_verifying(self):
         # While we're verifying, invalidations are ignored
-        wrapper, cache, loop, client, protocol, transport, send, respond = (
-            self.start())
+        wrapper, cache, loop, client, protocol, transport = self.start()
         protocol.data_received(sized(b'Z3101'))
         self.assertEqual(self.unsized(transport.pop(2)), b'Z3101')
-        self.assertEqual(self.parse(transport.pop()),
+        self.assertEqual(self.pop(),
                          [(1, False, 'register', ('TEST', False)),
                           (2, False, 'lastTransaction', ()),
                           ])
-        respond(1, None)
-        send('invalidateTransaction', b'b'*8, [b'1'*8])
-        self.assertFalse(wrapper.invalidateTransaction.called)
-        respond(2, b'a'*8)
-        send('invalidateTransaction', b'c'*8, [b'1'*8])
-        wrapper.invalidateTransaction.assert_called_with(b'c'*8, [b'1'*8])
-        wrapper.invalidateTransaction.reset_mock()
+        self.respond(1, None)
+        self.send('invalidateTransaction', b'b'*8, [b'1'*8], called=False)
+        self.respond(2, b'a'*8)
+        self.send('invalidateTransaction', b'c'*8, [b'1'*8], no_output=False)
+        self.assertEqual(self.pop(), (3, False, 'get_info', ()))
 
         # We'll disconnect:
         protocol.connection_lost(Exception("lost"))
@@ -535,17 +556,15 @@ class AsyncTests(setupstack.TestCase, ClientRunner):
 
         protocol.data_received(sized(b'Z3101'))
         self.assertEqual(self.unsized(transport.pop(2)), b'Z3101')
-        self.assertEqual(self.parse(transport.pop()),
+        self.assertEqual(self.pop(),
                          [(1, False, 'register', ('TEST', False)),
                           (2, False, 'lastTransaction', ()),
                           ])
-        respond(1, None)
-        send('invalidateTransaction', b'd'*8, [b'1'*8])
-        self.assertFalse(wrapper.invalidateTransaction.called)
-        respond(2, b'c'*8)
-        send('invalidateTransaction', b'e'*8, [b'1'*8])
-        wrapper.invalidateTransaction.assert_called_with(b'e'*8, [b'1'*8])
-        wrapper.invalidateTransaction.reset_mock()
+        self.respond(1, None)
+        self.send('invalidateTransaction', b'd'*8, [b'1'*8], called=False)
+        self.respond(2, b'c'*8)
+        self.send('invalidateTransaction', b'e'*8, [b'1'*8], no_output=False)
+        self.assertEqual(self.pop(), (3, False, 'get_info', ()))
 
     def test_flow_control(self):
         # When sending a lot of data (blobs), we don't want to fill up
@@ -553,8 +572,8 @@ class AsyncTests(setupstack.TestCase, ClientRunner):
         # seems a bit complicated. We'd rather pass an iterator that's
         # consumed as we can.
 
-        wrapper, cache, loop, client, protocol, transport, send, respond = (
-            self.start(finish_start=True))
+        wrapper, cache, loop, client, protocol, transport = self.start(
+            finish_start=True)
 
         # Give the transport a small capacity:
         transport.capacity = 2
@@ -564,52 +583,45 @@ class AsyncTests(setupstack.TestCase, ClientRunner):
         self.async('splat')
 
         # The first 2 were sent, but the remaining were queued.
-        self.assertEqual(self.parse(transport.pop()),
+        self.assertEqual(self.pop(),
                          [(0, True, 'foo', ()), (0, True, 'bar', ())])
 
         # But popping them allowed sending to resume:
-        self.assertEqual(self.parse(transport.pop()),
+        self.assertEqual(self.pop(),
                          [(0, True, 'baz', ()), (0, True, 'splat', ())])
 
         # This is especially handy with iterators:
         self.async_iter((name, ()) for name in 'abcde')
-        self.assertEqual(self.parse(transport.pop()),
-                         [(0, True, 'a', ()), (0, True, 'b', ())])
-        self.assertEqual(self.parse(transport.pop()),
-                         [(0, True, 'c', ()), (0, True, 'd', ())])
-        self.assertEqual(self.parse(transport.pop()),
-                         (0, True, 'e', ()))
-        self.assertEqual(self.parse(transport.pop()),
-                         [])
+        self.assertEqual(self.pop(), [(0, True, 'a', ()), (0, True, 'b', ())])
+        self.assertEqual(self.pop(), [(0, True, 'c', ()), (0, True, 'd', ())])
+        self.assertEqual(self.pop(), (0, True, 'e', ()))
+        self.assertEqual(self.pop(), [])
 
     def test_bad_protocol(self):
-        wrapper, cache, loop, client, protocol, transport, send, respond = (
-            self.start())
+        wrapper, cache, loop, client, protocol, transport = self.start()
         with mock.patch("ZEO.asyncio.client.logger.error") as error:
             self.assertFalse(error.called)
             protocol.data_received(sized(b'Z200'))
-            self.assert_(isinstance(error.call_args[0][1],
-                                    ZEO.Exceptions.ProtocolError))
+            self.assert_(isinstance(error.call_args[0][1], ProtocolError))
 
 
     def test_get_peername(self):
-        wrapper, cache, loop, client, protocol, transport, send, respond = (
-            self.start(finish_start=True))
+        wrapper, cache, loop, client, protocol, transport = self.start(
+            finish_start=True)
         self.assertEqual(client.get_peername(), '1.2.3.4')
 
     def test_call_async_from_same_thread(self):
         # There are a few (1?) cases where we call into client storage
         # where it needs to call back asyncronously. Because we're
         # calling from the same thread, we don't need to use a futurte.
-        wrapper, cache, loop, client, protocol, transport, send, respond = (
-            self.start(finish_start=True))
+        wrapper, cache, loop, client, protocol, transport = self.start(
+            finish_start=True)
 
         client.call_async_from_same_thread('foo', 1)
-        self.assertEqual(self.parse(transport.pop()), (0, True, 'foo', (1, )))
+        self.assertEqual(self.pop(), (0, True, 'foo', (1, )))
 
     def test_ClientDisconnected_on_call_timeout(self):
-        wrapper, cache, loop, client, protocol, transport, send, respond = (
-            self.start())
+        wrapper, cache, loop, client, protocol, transport = self.start()
         self.wait_for_result = super().wait_for_result
         self.assertRaises(ClientDisconnected, self.call, 'foo')
         client.ready = False
@@ -620,34 +632,35 @@ class AsyncTests(setupstack.TestCase, ClientRunner):
         # that caused it to fail badly if errors were raised while
         # handling data.
 
-        wrapper, cache, loop, client, protocol, transport, send, respond = (
-            self.start(finish_start=True))
+        wrapper, cache, loop, client, protocol, transport =self.start(
+            finish_start=True)
 
         wrapper.receiveBlobStart.side_effect = ValueError('test')
 
         chunk = 'x' * 99999
         try:
-            loop.protocol.data_received(b''.join((
-                sized(pickle.dumps(
-                    (0, True, 'receiveBlobStart', ('oid', 'serial')), 3)),
-                sized(pickle.dumps(
-                    (0, True, 'receiveBlobChunk',
-                     ('oid', 'serial', chunk)), 3)),
-                )))
+            loop.protocol.data_received(
+                sized(
+                    self.encode(0, True, 'receiveBlobStart', ('oid', 'serial'))
+                    ) +
+                sized(
+                    self.encode(
+                        0, True, 'receiveBlobChunk', ('oid', 'serial', chunk))
+                    )
+                )
         except ValueError:
             pass
-        loop.protocol.data_received(
-            sized(pickle.dumps(
-                (0, True, 'receiveBlobStop', ('oid', 'serial')), 3)),
-            )
+        loop.protocol.data_received(sized(
+            self.encode(0, True, 'receiveBlobStop', ('oid', 'serial'))
+            ))
         wrapper.receiveBlobChunk.assert_called_with('oid', 'serial', chunk)
         wrapper.receiveBlobStop.assert_called_with('oid', 'serial')
 
     def test_heartbeat(self):
         # Protocols run heartbeats on a configurable (sort of)
         # heartbeat interval, which defaults to every 60 seconds.
-        wrapper, cache, loop, client, protocol, transport, send, respond = (
-            self.start(finish_start=True))
+        wrapper, cache, loop, client, protocol, transport = self.start(
+            finish_start=True)
 
         delay, func, args, handle = loop.later.pop()
         self.assertEqual(
@@ -658,7 +671,7 @@ class AsyncTests(setupstack.TestCase, ClientRunner):
 
         # The heartbeat function sends heartbeat data and reschedules itself.
         func()
-        self.assertEqual(self.parse(transport.pop()), (-1, 0, '.reply', None))
+        self.assertEqual(self.pop(), (-1, 0, '.reply', None))
         self.assertTrue(protocol.heartbeat_handle != handle)
 
         delay, func, args, handle = loop.later.pop()
@@ -672,27 +685,6 @@ class AsyncTests(setupstack.TestCase, ClientRunner):
         protocol.connection_lost(None)
         self.assertTrue(handle.cancelled)
 
-    def unsized(self, data, unpickle=False):
-        result = []
-        while data:
-            size, message, *data = data
-            self.assertEqual(struct.unpack(">I", size)[0], len(message))
-            if unpickle:
-                message = pickle.loads(message)
-            result.append(message)
-
-        if len(result) == 1:
-            result = result[0]
-        return result
-
-    def parse(self, data):
-        return self.unsized(data, True)
-
-def response(*data):
-    return sized(pickle.dumps(data, 3))
-
-def sized(message):
-    return struct.pack(">I", len(message)) + message
 
 class MemoryCache:
 
@@ -745,6 +737,106 @@ class MemoryCache:
     def setLastTid(self, tid):
         self.last_tid = tid
 
+
+class ServerTests(Base, setupstack.TestCase):
+
+    # The server side of things is pretty simple compared to the
+    # client, because it's the clien't job to make and keep
+    # connections. Servers are pretty passive.
+
+    def connect(self, finish=False):
+        protocol = server_protocol()
+        self.loop = protocol.loop
+        self.target = protocol.zeo_storage
+        if finish:
+            self.assertEqual(self.pop(parse=False), best_protocol_version)
+            protocol.data_received(sized(b'Z4'))
+        return protocol
+
+    message_id = 0
+    target = None
+    def call(self, meth, *args, **kw):
+        if kw:
+            expect = kw.pop('expect', self)
+            target = kw.pop('target', self.target)
+            self.assertFalse(kw)
+
+            if target is not None:
+                target = getattr(target, meth)
+                if expect is not self:
+                    target.return_value = expect
+
+        self.message_id += 1
+        self.loop.protocol.data_received(
+            sized(self.encode(self.message_id, False, meth, args)))
+
+        if target is not None:
+            target.assert_called_once_with(*args)
+            target.reset_mock()
+
+        if expect is not self:
+            self.assertEqual(self.pop(),
+                             (self.message_id, False, '.reply', expect))
+
+    def testServerBasics(self):
+        # A simple listening thread accepts connections.  It creats
+        # asyncio connections by calling ZEO.asyncio.new_connection:
+        protocol = self.connect()
+        self.assertFalse(protocol.zeo_storage.notify_connected.called)
+
+        # The server sends it's protocol.
+        self.assertEqual(self.pop(parse=False), best_protocol_version)
+
+        # The client sends it's protocol:
+        protocol.data_received(sized(b'Z4'))
+
+        self.assertEqual(protocol.protocol_version, b'Z4')
+
+        protocol.zeo_storage.notify_connected.assert_called_once_with(protocol)
+
+        # The client registers:
+        self.call('register', False, expect=None)
+
+        # It does other things, like, send hearbeats:
+        protocol.data_received(sized(b'(J\xff\xff\xff\xffK\x00U\x06.replyNt.'))
+
+        # The client can make async calls:
+        self.send('register')
+
+        # Let's close the connection
+        self.assertFalse(protocol.zeo_storage.notify_disconnected.called)
+        protocol.connection_lost(None)
+        protocol.zeo_storage.notify_disconnected.assert_called_once_with()
+
+    def test_invalid_methods(self):
+        protocol = self.connect(True)
+        protocol.zeo_storage.notify_connected.assert_called_once_with(protocol)
+
+        # If we try to call a methid that isn't in the protocol's
+        # white list, it will disconnect:
+        self.assertFalse(protocol.loop.transport.closed)
+        self.call('foo', target=None)
+        self.assertTrue(protocol.loop.transport.closed)
+
+def server_protocol(zeo_storage=None,
+                    protocol_version=None,
+                    addr=('1.2.3.4', '42'),
+                    ):
+    if zeo_storage is None:
+        zeo_storage = mock.Mock()
+    loop = Loop()
+    sock = () # anything not None
+    new_connection(loop, addr, sock, zeo_storage)
+    if protocol_version:
+        loop.protocol.data_received(sized(protocol_version))
+    return loop.protocol
+
+def response(*data):
+    return sized(self.encode(*data))
+
+def sized(message):
+    return struct.pack(">I", len(message)) + message
+
 class Logging:
 
     def __init__(self, level=logging.ERROR):
@@ -762,5 +854,6 @@ class Logging:
 
 def test_suite():
     suite = unittest.TestSuite()
-    suite.addTest(unittest.makeSuite(AsyncTests))
+    suite.addTest(unittest.makeSuite(ClientTests))
+    suite.addTest(unittest.makeSuite(ServerTests))
     return suite